Event Posters
Can now add and view event posters. Pictures can be selected from the user device and they are also uploaded to firebase storage when the event is created.

There is no default image set so for now you have to click the blank space beneath "Guests Allowed" when creating an event.